+++ src/solv_simple/fn_smurf/fn_smurf.cc
@@ -15,13 +15,13 @@ void PrintSmurfStateEntry(void *pState) {
    SmurfStateEntry *pSmurfState = (SmurfStateEntry *)pState;
 	d9_printf9("SM Var=%d, v=T:%x, v=F:%x, TWght=%4.6f, FWght=%4.6f, NextGT=%x, NextLT=%x, Next=%x\n",
               pSmurfState->nTransitionVar, 
-              (unsigned int)pSmurfState->pVarIsTrueTransition,
-              (unsigned int)pSmurfState->pVarIsFalseTransition,
+              (uintptr_t)pSmurfState->pVarIsTrueTransition,
+              (uintptr_t)pSmurfState->pVarIsFalseTransition,
               pSmurfState->fHeurWghtofTrueTransition,
               pSmurfState->fHeurWghtofFalseTransition,
-              (unsigned int)pSmurfState->pNextVarInThisStateGT,
-              (unsigned int)pSmurfState->pNextVarInThisStateLT,
-              (unsigned int)pSmurfState->pNextVarInThisState);
+              (uintptr_t)pSmurfState->pNextVarInThisStateGT,
+              (uintptr_t)pSmurfState->pNextVarInThisStateLT,
+              (uintptr_t)pSmurfState->pNextVarInThisState);
 }
 
 void PrintSmurfStateEntry_dot(void *pState) {
@@ -29,12 +29,12 @@ void PrintSmurfStateEntry_dot(void *pState) {
    SmurfStateEntry *pTempState = pSmurfState;
 	while(pTempState != NULL) {
 		fprintf(stdout, " b%x->b%x [style=solid,fontname=\"Helvetica\",label=\"%s\"]\n", 
-              (unsigned int)pState, 
-              (unsigned int)pTempState->pVarIsTrueTransition, 
+              (uintptr_t)pState, 
+              (uintptr_t)pTempState->pVarIsTrueTransition, 
               s_name(arrSimpleSolver2IteVarMap[pTempState->nTransitionVar]));
 		fprintf(stdout, " b%x->b%x [style=dotted,fontname=\"Helvetica\",label=\"-%s\"]\n",
-              (unsigned int)pState,
-              (unsigned int)pTempState->pVarIsFalseTransition,
+              (uintptr_t)pState,
+              (uintptr_t)pTempState->pVarIsFalseTransition,
               s_name(arrSimpleSolver2IteVarMap[pTempState->nTransitionVar]));
 //		fprintf(stdout, "%x [
 		pTempState = (SmurfStateEntry *)pTempState->pNextVarInThisState;
@@ -45,7 +45,7 @@ void PrintInferenceStateEntry(void *pState) {
    InferenceStateEntry *pInferenceState = (InferenceStateEntry *)pState;
 	d9_printf4("IN Var=%d, Inf=%x, Polarity=%d\n",
               pInferenceState->nTransitionVar,
-              (unsigned int)pInferenceState->pVarTransition,
+              (uintptr_t)pInferenceState->pVarTransition,
               pInferenceState->bPolarity);
 }
 
@@ -53,15 +53,15 @@ void PrintInferenceStateEntry_dot(void *pState) {
    InferenceStateEntry *pInferenceState = (InferenceStateEntry *)pState;
 	if(pInferenceState->bPolarity)
 	  fprintf(stdout, " b%x->b%x [style=solid,fontname=\"Helvetica\",label=\"%s\"]\n", 
-             (unsigned int)pInferenceState, 
-             (unsigned int)pInferenceState->pVarTransition, 
+             (uintptr_t)pInferenceState, 
+             (uintptr_t)pInferenceState->pVarTransition, 
              s_name(arrSimpleSolver2IteVarMap[pInferenceState->nTransitionVar]));
 	else
 	  fprintf(stdout, " b%x->b%x [style=dotted,fontname=\"Helvetica\",label=\"-%s\"]\n", 
-             (unsigned int)pInferenceState, 
-             (unsigned int)pInferenceState->pVarTransition, 
+             (uintptr_t)pInferenceState, 
+             (uintptr_t)pInferenceState->pVarTransition, 
              s_name(arrSimpleSolver2IteVarMap[pInferenceState->nTransitionVar]));
-	fprintf(stdout, " b%x [shape=\"ellipse\", label=\"I\"]\n", (unsigned int)pInferenceState);
+	fprintf(stdout, " b%x [shape=\"ellipse\", label=\"I\"]\n", (uintptr_t)pInferenceState);
 }
 
 void FreeSmurfStateEntry(void *pState) {
