+++ src/solv_simple/solv_simple_init.cc
@@ -529,7 +529,7 @@ void PrintSmurfs(BDDNode **bdds, int size) {
 		} else {
 			fprintf(stdout, "digraph Smurf {\n");
 			fprintf(stdout, " graph [concentrate=true, nodesep=\"0.30\", ordering=in, rankdir=TB, ranksep=\"2.25\"];\n");
-			fprintf(stdout, " b%x [shape=box fontname=""Helvetica"",label=""T""];\n", (unsigned int)pTrueSimpleSmurfState);
+			fprintf(stdout, " b%x [shape=box fontname=""Helvetica"",label=""T""];\n", (uintptr_t)pTrueSimpleSmurfState);
 			PrintSmurf_dot(ReadSmurfStateIntoTable(bdds[i], NULL, 0));
 			fprintf(stdout, "}\n");
 		}
